Fix clang-tidy ci test (#7)
Fixes a bunch of issues reported by clang-tidy.
Mostly mechanical, except for addition of an interface for the Uri class.

Previously, /source/common/utility.h could be found by clang-tidy while in here,
with nighthawk as a root repository, it can't. That couldn't be fixed without adding a cyclic
dependency. (It's nice to have regardless).

With this included, we're close to having a clean clang-tidy report.

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>