Experiment: Mixer integration server for testing
Putting this out here for discussion:
This PR absorbs int_server.cc/int_server.h from
https://github.com/duderino/proxy/blob/jblatt_mixer_fault_testing/test/integration/

Then it goes on to update the end-to-end integration test for the NH client
to rely on this server for testing, swapping out Envoy's integration
server. This allows us to ditch the awkward fork() we have to do
to avoid runtime conflicts and deflake CI.

Also enhanced and added to the existing client tests while playing around with
the new server, to get a feel for ease of use (which seems to be pretty easy).

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>