Back out using `oneoff`, add TODO to fix overriding with default value. (#63)
This disables HttpTestServerIntegrationTest.TestZeroLengthRequest and backs out the oneoff proto hack we used to make overriding with 0 work.

The now disabled tests attempts to merge {response_size:0} into the default configuration, which will not take take effect, because the proto merge ignores the type-default (0) when merging. Conceptually proto doesn't make a distinction between "default" and "not specified". Tihs also adds a TODO to provide a way around this.

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>