Use MilestoneTracker in StreamDecoder
Replaces the latency measurement code in StreamDecoder
with the MilestoneTracker. Adds a couple of measurement
points to the transaction.

Also adds a test to verify StreamDecoder latency is measured as expected
using SimulatedTime.

The api of the MilestoneTracker can be improved, and we can now output
more histograms. Due in a follow-up.
is next up.

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>