Service: address back to back execution race
Address a race discovered via repeated parallel execution of
`ServiceTest, BackToBackExecution` by adding a short timeout
before concluding that there's still a previous future active and handling
the earlier test.

Theoretically it's still possible to run into this, but  in practice I think this
is fair enough for actual usage, and it is sufficient to deflake the
test which detected it on my machine.

Thanks @htuch for discovering the issue

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>