[WIP] Add request generator concept
This adds some infrastructure to allow for dynamic request generation.

This first step is a no-op, in a sense that current functionality
is left unchanged, as it wires in a StaticHeaderGeneratorImpl,
which is rather boring.

However, for 0.3 we'd like to have a repay capability, at which
point this concept can be re-used to plug in a different type
of request generator, which translates yields headers to NH
from a grpc stream. Another feature request has been to allow
for request distributions; this lowers the bar for that considerably.

Related:
- https://github.com/envoyproxy/nighthawk/issues/34
- https://github.com/envoyproxy/nighthawk/issues/102

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>