Tracking per-worker counter accumulations (#116)
Having the per-worker counter accumulations available is useful because it allows us to observe/detect certain types of inter-worker imbalances. In 0.2 we had this feature but after switching to tls-statsistics on master we lost that capability. This restores that feature by creating distinct clusters for each worker, thereby isolating statistics for them again. At the end of a test run the per-worker counters are aggregated in the main thread to get the global values.

Fixes https://github.com/envoyproxy/nighthawk/issues/111

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>