Phases concept
Introduces a concept `Phases` into the system.
Phases represent pieces of the overall execution, scoping
statistics, configuration, and termination predicates.

For example, a warmup phase could be statically added based on
a clone of the main phase with a rate limiter which smoothly
ramps up to the target qps.

Currently a no-op, meant to initiate discussion and
have something concrete to start with.

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>