Optionize termination predicates (#176)
Allows configuration of termination & failure predicates.

Sample CLI usage:

# Stop after 10 requests, indicating success
--termination-predicate "benchmark.http_2xx: 10"
# Fail after a single connection failure
--failure-predicate "upstream_cx_connect_fail: 0"
Adds tracking and publishing of the effective execution duration. As we are now able
to terminate before the specified execution duration.

Note: renames the worker tag to cluster so we share the same counter namespace. This allows the predicates to query the involved counters without going through hoops. As this is change does not propagate to the output, this is for internal handling only, and as such it is not a breaking change.

Together with #167, fixes #154

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>