StreamDecoder: amend onXXXWriteBufferHighWatermark
Following a report that came in via a backchannel,  remove
NOT_IMPLEMENTED_GCOVR_EXCL_LINE and add tests that will hit those.

Suspiciously, server side expectations fail with larger request body
sizes. That is pending investigation, I will file a separate issue for
that. Meantime, the test lines for that are commented out.

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>