Optimization: static request body
We always send the same request body today.
It is possible to optimize that case and reduce copies.
This greatly speeds up achieved transfer rates as well as allows
much higher concurrencies to be used when testing uploads.
(This was initiated based on a field report containing memory
exhaustion symptoms)

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>