RemoteHeaderSource & stub grpc service HeaderStreamRequest (#137)
Adds a `--request-source` option, which directs Nighhtawk's workers to connect to a remote grpc service to pull to-be-replayed request headers from. These headers will be merged on the client side with any uri configuration specified on the client side.

```
--request-source <uri format>
Remote gRPC source that will deliver to-be-replayed traffic. Each
worker will separately connect to this source. For example
grpc://127.0.0.1:8443/.
```

Summary of introduced components:
- Introduces a grpc client that NH's workers rely on to talk to a service to get to-be-replayed contents
- Introduces a new `RequestSource` implementation which uses the above to source remote request data.
- Introduces a new dummy grpc service method which implements a stub header stream for testing
- Introduces a python wrapper for our grpc service for use in integration testing
- Makes our grpc service listen to certain signals so we can shut it down nicely from the integration tests.

Related issues:
- https://github.com/envoyproxy/nighthawk/issues/34
- https://github.com/envoyproxy/nighthawk/issues/102

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>