Simple warmup: allow opting out
The current single request we issue for the simple warmup pollutes
some of the counters we emit in our output.

Looking forward, when landing generic phases, we might want to
default to opting out.
This new flag could then serve as an opt-in afterwards, for those
who want to preserve old behavior.

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>