Dead code: eliminate StoreFactory
Just noticed some code that no longer is actually used.

Context:

StoreFactory became obsolete when it became apparent that
the isolated stats store implementation didn't have enough
production mileage to be a suitable candidate. We switched
to use the thread local backing store.

Today NH is basically hard-coupled to that store, as it builds
on top of some of its shared semantics, so this abstraction
probably will not be useful in the foreseeable future.

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>