Update Envoy to 888e0e28900a470df448c65d7b99d8065fd60251 (#331)
This Envoy update is a bit more involved, because it forced moving away from the
legacy connection pools it had, which got eliminated from the code base. So we migrate
to Envoy's new pool models.

* The h1 pool migrates seamlessly to the new model with minor changes.

* The experimental h2 connection pool that supports multiple connections gets dropped. We now offer multiple connection support based on the new H2 pool. Behavior has changed, but there was a warning about that in the description of the --experimental flag (see changes to the test associated to our earlier experimental h2
pool)

It should now be possible to add new options to add better support for h2 & multiple connections,
and deprecate or remove the --experimental flag for that.

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>
Signed-off-by: qqustc <qqin@google.com>