Update Envoy to 08464ecdc0c93846f3d039d0f0c6fed935f5bdc8
Note: adds a patch for Envoy that allows us to re-use its
linkstamping mechanism, to avoid an undefined reference
when building our artifacts.

We need this, because we pull in the versioning mechanism,
but the actual linkstamping implementation used is hidden
from us in our build. The patch creates a visible copy
of what we need.

This all isn't ideal. Maybe we can add a version to Envoy's
build system that we actually can re-use, or change its
dependencies so that we can consume what we need without
pulling in the header that rely on the linkstamp
implementation.

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>