Adaptive Load exponential search step controller (#460)
A StepController plugin that searches exponentially for a bad input that causes metrics to go outside thresholds, then performs a binary search between the last known good input and the bad input. It should not exceed 2x the healthy capacity of the system under test. 

The exponent of the exponential search can be adjusted from 2. 

Note that this algorithm does not take into account the magnitude of the metric score, only the sign.

This PR also adds FakeInputVariableSetter in order to use it in the unit tests for the StepController, and at the same time removes  multiple implementation inheritance for all InputVariableSetters.