CI: Track integration test coverage separately from unit tests (#481)
Following an offline discussion about coverage flakes in CI, and how to best
improve there we decided to exclude the integration tests altogether from strict coverage
treshold expectations. As such, we lower the unit testing coverage threshold to 92.1%
here. We still run the integration tests in a new CI task, with the intent to make that not
required to pass for merging, so informational only.

Fixes #362 

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>