Echo back and track origin request-receipt timing deltas (#477)
This adds a rough cut of the capability to the test server to allow tracking of the delta's between
request arrival times, and echo these back in a configurable response header name.

Subsequently, the nighthawk_client gets a feature to track numbers received via an arbitrary
header in a histogram. Right now it's hard coded to track `x-nh-do-not-use-origin-timings`. 
But the plan would be to allow for at least one configurable header name to track. 

Combined, these test server and client features enable obtaining insight into the origin-side point of
view regarding arrival timings [1]. The direct use case is manual inspection / sanity checking, but possibly
divergence of what would be expected based on client side configuration can be used as a way to quantify
"distortion" introduced by an intermediary proxy.

In a follow up, documentation, configuration of the client to specify the response header name that should
be used to track latency, and end to end tests will be added. 

When setting up a test between the Nighthawk client and test server using this feature, a new histogram
will be send to the output. On my machine, a simple 1000qps test yields the following perspective from
the test-server when it comes to deltas between inbound requests:

```
  0.5         2519        0s 000ms 999us 
  0.75        3767        0s 001ms 001us 
  0.8         4002        0s 001ms 002us 
  0.9         4508        0s 001ms 003us 
  0.95        4756        0s 001ms 004us 
  0.990625    4953        0s 001ms 005us 
  0.99902344  4995        0s 001ms 015us ```

[1] Concretely, if a straightforward 1000 qps is configured, the histogram obtained from tracking the server side should reasonably reflect that and show a histogram diverging very little from 1ms.