Set explicit --jobs value in ci/do_ci.sh only when running on CircleCI (#506)
For local `ci/do_ci.sh` runs, this change allows the explicit `--jobs` setting from `bazelrc` files to take effect, or falls back on the default `--jobs=auto`.

This change stops passing ` --jobs $NUM_CPUS` for non-CircleCI Bazel runs, but any other uses of `NUM_CPUS` are assumed to be critical and are all left unchanged. For example, `NUM_CPUS` must be in the environment for the call out to `ci/run_clang_tidy.sh` to work. For clang-tidy and anything else that doesn't use a remote Bazel build, we should always allow `NUM_CPUS` to take effect, both locally and on CircleCI.

Motivation for this change:
When using a `user.bazelrc` that uses Remote Build Execution with a high `--jobs` value, we discovered that `NUM_CPUS` in `ci/do_ci.sh` is artificially limiting the parallelism of the remote builds based on the local CPU count.

(`--jobs` is only useful for limiting the amount of parallelism, not increasing it: https://docs.bazel.build/versions/master/user-manual.html#flag--jobs.)

Also fixing the printing of the help message when the user forgot the arg.