Add FilterConfigurationBase
Base class for test server http filter extension configuration.
Provides support for shared configuration and consolidates
shared code.

Potentially useful in that it provides a path to cache parsed
and interpreted configuration accross extensions as it flows
through the server pipeline. This isn't a goal for now however,
this is a prelude to #512, which includes it and provides a means
to see the end goal.

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>