Fix jitter behavior for large values. (#552)
By chance weird behaviour was observed for when specifying large jitter
values. This warranted analysis, and upon investigation the root problem seems
to be that when configuring `--uniform-jitter` to a value that would allow the
adjusted value(s) to overlap with the pacing of the underlying rate limiter,
the new overlapping value could overwrite the old one that was scheduled.

This modifies behaviour and tests to fix and consolidate, by modifying 
`DistributionSamplingRateLimiter` so that it can queue adjusted release
timings instead of storing a single one. This will allow it to accumulate multiple
release timings that are adjusted for jitter. 
This ensures that the pacing of the rate limiter that modifies timings to add jitter
gets properly disconnected from the pacing of the underlying rate limiter.

NOTE: There are no known consumers running into problems / inaccuracies
because of this. To tease the problem out, either one needs to:
- configure `--jitter-uniform` in a way that it would adjust timings to overlap
with the frequency imposed by `--rps`, or
- configure any combination of `--jitter-uniform > 0s` and `--burst-size > 0`.

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>