Use system time for scheduling. (#549)
Small refactor that makes the client rely on system time instead of monotonic time
for scheduling the start of worker executions. System clocks can be synchronized
across machines, and this may come in handy when we start facilitating horizontal
scaling. 

Note: `SequencerImpl` gets modified to re-use the execution duration that the `RateLimiter`
it uses already tracks, in favour of its own tracking. This is a small clean up.

Apart from the actual switching from monotonic time to wall clock time, this should be a
mechanical change. 

This change will make things easier if we would like to add an option to schedule the time at
which an execution will start. 
This in turn could be useful when directing clients running on multiple machines to start, as a
means to have them start at approximately the same time. 
(the approximation would mostly depend on how well the wall clock time is synchronised across
machines that are involved).

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>