CI: Fix broken docker tasks
I think the root cause is that the alpine-glibc image updates,
and through that dependency requirements can no longer be met.

This PR
- pins the alpine image at the latest version
- changes the dependency versions so they work again

Note: the `>=` for python was changed to `=` as otherwise we'd
write a file to disk mentioning the version, this seemed contra
intent.

There's a complaint [1] emitted during the image build, so we need to
see if everything is OK when CI finishes.

[1] /usr/glibc-compat/sbin/ldconfig: /usr/glibc-compat/lib/ld-linux-x86-64.so.2 is not a symbolic link

Signed-off-by: Otto van der Schaaf <ovanders@redhat.com>