Examples: a place to stash practical use cases (#637)
This captures something I used and for which I had to dig a bit in
the codebase to get it set up. Stashing it here may help others and
lower the bar for consumption.

I anticipate there may be more examples like this that we could
add over time. So ideally this PR will be assessed both on structure
of the initiative (/examples) and contents (MULTIPLE_ENDPOINTS.md)

Signed-off-by: Otto van der Schaaf <ovanders@redhat.com>