Remove premature optimization to make request source factory more reusable (#703)
Fixing issue where nighthawk makes request source hard to reuse due to caching the options_list and removing associated todo. Caching a single list in the factory was a premature optimization and resulted in nighthawk being hard to reuse. Also changing so that an empty optionslist will return the default header instead of throwing an error. Added associated test for loading two request sources with the same factory.

Signed-off-by: William Juan <66322422+wjuan-AFK@users.noreply.github.com>