Avoid compare Protobuf TextFormat with Protobuf DebugString (#778)
The Protobuf DebugString format is an unstable format and it may be different from the Protobuf TextFormat at some point in the future. To make this unit test more robust,  
we deserialize the text proto (in TextFormat) from the golden file first, and then compare the DebugString of that deserialized proto with the actual value (a DebugString).