Increase hard shutdown timeout to 30 seconds (#1193)
This is inline with the default connection connect timeout period in 30 seconds. For well behaved clients with short response latencies, this should be a no-op as we will end prematurely once the pool is idle. This does allow for clients with high latencies a long enough time to resolve gracefully.

Signed-off-by: Tom Zhang <4367421+tomjzzhang@users.noreply.github.com>