Update Nighthawk Python dependencies (Aug 29, 2024) (#1215)
Update Nighthawk Python dependencies, pinning setuptools<=70.3.0.

Due to Bazel not supporting spaces in filenames, we have to pin setuptools to a version earlier than 71.0.0, which moves away from vendorized packages towards naturally installed ones. This eliminated a workaround which unintentionally allowed  Bazel to use the jaraco.text package despite it using a filename with a space in it (which usually doesn't play nice with Bazel): https://github.com/pypa/setuptools/issues/4487#issuecomment-2259039157.

We can remove this pin when Envoy (and therefore Nighthawk) is migrated to build with Bazel 7. This will allow us to build with an experimental flag that allows spaces in filenames: https://bazel.build/reference/command-line-reference#flag--experimental_inprocess_symlink_creation.

Signed-off-by: Sebastian Avila <sebastianavila@google.com>